minimize static bounds
