+++ node/service/src/lib.rs
@@ -21,7 +21,7 @@
 pub mod chain_spec;
 mod grandpa_support;
 mod parachains_db;
-pub(crate) mod relay_chain_selection;
+mod relay_chain_selection;
 
 #[cfg(feature = "full-node")]
 mod overseer;
@@ -49,7 +49,6 @@ use {
 	sp_trie::PrefixedMemoryDB,
 	sc_client_api::ExecutorProvider,
 	grandpa::{self, FinalityProofProvider as GrandpaFinalityProofProvider},
-	sp_runtime::traits::Header as HeaderT,
 };
 
 #[cfg(feature = "full-node")]
@@ -101,7 +100,18 @@ pub use service::{
 };
 pub use service::config::{DatabaseConfig, PrometheusConfig};
 pub use sp_api::{ApiRef, Core as CoreApi, ConstructRuntimeApi, ProvideRuntimeApi, StateBackend};
-pub use sp_runtime::traits::{DigestFor, HashFor, NumberFor, Block as BlockT, self as runtime_traits, BlakeTwo256};
+pub use sp_runtime::{
+	generic,
+	traits::{
+		DigestFor,
+		HashFor,
+		NumberFor,
+		Block as BlockT,
+		Header as HeaderT,
+		self as runtime_traits,
+		BlakeTwo256,
+	},
+};
 
 #[cfg(feature = "kusama-native")]
 pub use kusama_runtime;
@@ -114,6 +124,60 @@ pub use westend_runtime;
 /// The maximum number of active leaves we forward to the [`Overseer`] on startup.
 #[cfg(any(test,feature = "full-node"))]
 const MAX_ACTIVE_LEAVES: usize = 4;
+
+/// Provides the header and block number for a hash.
+///
+/// Decouples `sc_client_api::Backend` and `sp_blockchain::HeaderBackend`.
+pub trait HeaderProvider<Block, Error=sp_blockchain::Error>: Send + Sync + 'static
+where
+	Block: BlockT,
+	Error: std::fmt::Debug + Send + Sync + 'static,
+{
+	/// Obtain the header for a hash.
+	fn header(&self, hash: <Block as BlockT>::Hash) -> Result<Option<<Block as BlockT>::Header>, Error>;
+	/// Obtain the block number for a hash.
+	fn number(&self, hash: <Block as BlockT>::Hash) -> Result<Option<<<Block as BlockT>::Header as HeaderT>::Number>, Error>;
+}
+
+impl<Block, T> HeaderProvider<Block> for T
+where
+	Block: BlockT,
+	T: sp_blockchain::HeaderBackend<Block> + 'static,
+{
+	fn header(&self, hash: Block::Hash) -> sp_blockchain::Result<Option<<Block as BlockT>::Header>> {
+		<Self as sp_blockchain::HeaderBackend<Block>>::header(self, generic::BlockId::<Block>::Hash(hash))
+	}
+	fn number(&self, hash: Block::Hash) -> sp_blockchain::Result<Option<<<Block as BlockT>::Header as HeaderT>::Number>> {
+		<Self as sp_blockchain::HeaderBackend<Block>>::number(self, hash)
+	}
+}
+
+/// Decoupling the provider.
+///
+/// Mandated since `trait HeaderProvider` can only be
+/// implemented once for a generic `T`.
+pub trait HeaderProviderProvider<Block>: Send + Sync + 'static
+where
+	Block: BlockT,
+{
+	type Provider: HeaderProvider<Block> + 'static;
+
+	fn header_provider(&self) -> &Self::Provider;
+}
+
+
+impl<Block, T> HeaderProviderProvider<Block> for T
+where
+ 	Block: BlockT,
+ 	T: sc_client_api::Backend<Block> + 'static,
+ {
+ 	type Provider = <T as sc_client_api::Backend<Block>>::Blockchain;
+
+ 	fn header_provider(&self) -> &Self::Provider {
+		self.blockchain()
+ 	}
+}
+
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
@@ -500,7 +564,7 @@ where
 		.unwrap_or_default()
 		.into_iter()
 		.filter_map(|hash| {
-			let number = client.number(hash).ok()??;
+			let number = HeaderProvider::number(client, hash).ok()??;
 
 			// Only consider leaves that are in maximum an uncle of the best block.
 			if number < best_block.number().saturating_sub(1) {
